+++ Projects/App/Sources/Generic/AppFlow+MovingFlow.swift
@@ -12,14 +12,11 @@ import UIKit
 
 struct MovingFlow {
     @Inject var client: ApolloClient
-    @Cached var moveData: Data?
 }
 
 extension MovingFlow: Presentable {
     public func materialize() -> (UIViewController, Disposable) {
         let viewController = UIViewController()
-
-        if let moveData = moveData {}
 
         return (viewController, DisposeBag())
     }
